refactor(core): refactor `jest` migration
